Use styled components instead of CSS
While implementing this a lot of the flaws of the MainPage file have
been highlighted. `Thought` needs pulling out into it's own file, as
does just about everything rendered as a child of it. This would help us
to clean up quite a lot. I think I'd like to go through the whole app
and do a general refactor/tidy up before taking a look at adding
styled-components again. There's quite a bit to be done to bring it to a
place I'd be comfortable with moving forward, otherwise refactoring
after adding styled-components is going to be very painful.

All that aside, it was actually really just as easy as I figured it
would be, add the dependency and you're good to go. Overall a pleasant
experience. I like that it's no more complicated than that and there's
no reason we couldn't use styled-components in this project.