Create reusable Tag component
There appears to be three separate implementations of the tag component
used in the app. In the interest of cleaning up the code I've created a
reusable Tag component and plugged it into those three places.

I've done this with #245 in mind. We've got some helpers at the top of
the `Tag.js` file to handle styling, but `styled-components` will be
able to handle all of this inline, so this should clean up significantly
once that refactor takes place.

I've also decided to render just a `Link` instead of using a `div`. This
is simply to reduce complexity in this component. I originally wrote
this with a conditional to check if `href` existed, and then render a
`Link` if it did, otherwise to render a `div` and it was a bit of a
dog's dinner. This solution is much cleaner than the alternative.