Create a seperate runSetup function
This does the `npm install` (if required) and the database setup.

Since this is now happening in the generation phase, this means we don't need
`compileAndSetup` anymore.

For clarity and brevity, rename PackageJson to Npm throughout. So it's
`NpmDependencies`

Since compile now sets up the dependencies, we only need to compile in any db command