Fix error Shayne pointed out with rendering of failures
I've now expanded Failure to take a title and a main message. The command processor renders this
accurately now. I've gotten rid of manual Failure messages in the setup phase, as it was already
redundant and doesn't fit this pattern really well. We may want to rethink this a bit more in the future,
as it would be useful to know what "phase" we are in, now only expressed in failure titles.

Rename Notification to Info